+++ controller/konnect/index_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/kong/gateway-operator/controller/konnect/constraints"
 	"github.com/kong/gateway-operator/modules/manager/scheme"
 
+	commonv1alpha1 "github.com/kong/kubernetes-configuration/api/common/v1alpha1"
 	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
@@ -45,7 +46,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongServiceSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name: "cp-1",
@@ -63,7 +64,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongServiceSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name:      "cp-1",
@@ -82,7 +83,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongServiceSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name:      "cp-1",
@@ -112,7 +113,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongRouteSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name: "cp-1",
@@ -130,7 +131,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongRouteSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name:      "cp-1",
@@ -149,7 +150,7 @@ func TestIndexKonnectGatewayControlPlaneRef(t *testing.T) {
 						Name:      "obj1",
 					},
 					Spec: configurationv1alpha1.KongRouteSpec{
-						ControlPlaneRef: &configurationv1alpha1.ControlPlaneRef{
+						ControlPlaneRef: &commonv1alpha1.ControlPlaneRef{
 							Type: configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef,
 							KonnectNamespacedRef: &configurationv1alpha1.KonnectNamespacedRef{
 								Name:      "cp-1",