+++ controller/konnect/ops/server_url_test.go
@@ -6,6 +6,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/kong/gateway-operator/controller/konnect/ops"
+
+	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
 
 func TestServerURL(t *testing.T) {
@@ -27,8 +29,33 @@ func TestServerURL(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			got := ops.NewServerURL(tc.input)
+			got := ops.NewServerURL[konnectv1alpha1.KonnectGatewayControlPlane](tc.input)
 			require.Equal(t, tc.expected, got.String())
 		})
 	}
+
+	t.Run("KonnectCloudGatewayNetwork", func(t *testing.T) {
+		konnectTestCases := []struct {
+			name     string
+			input    string
+			expected string
+		}{
+			{
+				name:     "us",
+				input:    "us.api.konghq.com",
+				expected: "https://global.api.konghq.com",
+			},
+			{
+				name:     "eu",
+				input:    "eu.api.konghq.com",
+				expected: "https://global.api.konghq.com",
+			},
+		}
+		for _, tc := range konnectTestCases {
+			t.Run(tc.name, func(t *testing.T) {
+				got := ops.NewServerURL[konnectv1alpha1.KonnectCloudGatewayNetwork](tc.input)
+				require.Equal(t, tc.expected, got.String())
+			})
+		}
+	})
 }