+++ controller/konnect/watch_kongcacertificate.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/kong/gateway-operator/modules/manager/logging"
 
+	commonv1alpha1 "github.com/kong/kubernetes-configuration/api/common/v1alpha1"
 	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
@@ -72,7 +73,7 @@ func enqueueKongCACertificateForKonnectAPIAuthConfiguration(cl client.Client) ha
 			}
 
 			switch cpRef.Type {
-			case configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef:
+			case commonv1alpha1.ControlPlaneRefKonnectNamespacedRef:
 				nn := types.NamespacedName{
 					Name:      cpRef.KonnectNamespacedRef.Name,
 					Namespace: cert.Namespace,
@@ -103,7 +104,7 @@ func enqueueKongCACertificateForKonnectAPIAuthConfiguration(cl client.Client) ha
 					},
 				})
 
-			case configurationv1alpha1.ControlPlaneRefKonnectID:
+			case commonv1alpha1.ControlPlaneRefKonnectID:
 				ctrllog.FromContext(ctx).Error(
 					fmt.Errorf("unimplemented ControlPlaneRef type %q", cpRef.Type),
 					"unimplemented ControlPlaneRef for KongCACertificate",